Enforce int output for DNDarray.size incl. when DNDarray.shape is `()`

+++ heat/core/dndarray.py
@@ -191,7 +191,9 @@ class DNDarray:
         """
         Number of total elements of the ``DNDarray``
         """
-        return torch.prod(torch.tensor(self.gshape, device=self.device.torch_device)).item()
+        return torch.prod(
+            torch.tensor(self.gshape, dtype=torch.int, device=self.device.torch_device)
+        ).item()
 
     @property
     def gnbytes(self) -> int:
@@ -1617,7 +1619,7 @@ class DNDarray:
         Return a 1-element `torch.Tensor` strided as the global `self` shape.
         Used internally for sanitation purposes.
         """
-        return torch.ones((1,)).as_strided(self.gshape, [0] * self.ndim)
+        return torch.ones((1,), dtype=torch.int8).as_strided(self.gshape, [0] * self.ndim)
 
     @staticmethod
     def __xitem_get_key_start_stop(
